Updating Installation Instructions
Composer is now able to select the latest compatible release when a library is requested via `composer require`, this is the recommended way of installing new packages. Editing the json file is not recommended and may cause other issues if a bad update/install is executed afterwards. `require` is an all in one solution that completes the whole process with less chance for mistakes.